Mario's Changes
So, I forgot I was on Karington's branch, and made my changes.  I got the nav bar to be an include deal.  There is a div for the navbar and a div for the footer.  The files that hold the actual stuff is in the "shared" folder.  Should be simple to figure out.  I added basic bootstrap stuff to everyone's pages.  Here's to hoping it doesn't suck.